Add logger
User should be able to access logging/debug info when exceptions are caught

[#27334181]

This is not quite ready for prime-time. Gonna need to do some more research on best-practices for implementing logging in a gem.

It 'works for me' but not confident in what it'll do in the wild.

Plus: I need to figure out how to make it suppress logging in spec tests.