Fix bug with NetBible
giving an error because I wasn't referencing the Helper::Text on the call to clean_verse_start/end.
For some reason this was passing the test but not working outside. Still need to debug that but will ship updated version first to fix the problem as NetBible is broken as-is.